# api/openapi.yaml
openapi: 3.0.1
info:
  title: eMekdep API
  description: ''
  version: 1.0.0
...
paths:
...
  /teacher/timetable:
    get:
      summary: /teacher/timetable
      deprecated: false
      description: ''
      tags:
        - Journal
      parameters:
        - name: date
          in: query
          description: ''
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - *ref_5
                  - $ref: '#/components/schemas/TeacherTimetable'
      security: []
...
  /subjects:
    get:
      summary: /subjects
      deprecated: false
      description: ''
      tags:
        - Classroom
      parameters:
        - name: classroom_ids
          in: query
          description: ''
          required: false
          schema:
            type: array
            items:
              type: integer
        - name: subject_names
          in: query
          description: ''
          required: false
          schema:
            type: array
            items:
              type: string
        - name: teacher_ids
          in: query
          description: ''
          required: false
          schema:
            type: array
            items:
              type: integer
        - name: is_second_teacher
          in: query
          description: ''
          required: false
          schema:
            type: boolean
        - name: sort
          in: query
          description: ''
          required: false
          schema:
            type: string
            enum:
              - id
              - classroom_id
              - subject_id
              - teacher_id
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - *ref_5
                  - properties:
                      data:
                        type: object
                        properties:
                          items:
                            type: array
                            items: &ref_54
                              $ref: '#/components/schemas/SubjectList'
                    type: object
      security: []
    post:
      summary: /subjects
      deprecated: false
      description: ''
      tags:
        - Classroom
      parameters: []
      requestBody:
        content:
          application/json:
            schema: &ref_26
              $ref: '#/components/schemas/SubjectCreate'
            example: ''
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - *ref_5
                  - properties:
                      data:
                        type: object
                        properties:
                          item: &ref_25
                            $ref: '#/components/schemas/Subject'
                    type: object
      security: []
    delete:
      summary: /subjects
      deprecated: false
      description: ''
      tags:
        - Classroom
      parameters:
        - name: ids
          in: query
          description: ''
          required: false
          schema:
            type: array
            items:
              type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - *ref_5
                  - properties:
                      data:
                        type: object
                        properties:
                          classrooms:
                            type: array
                            items: *ref_25
                    type: object
      security: []
  /subjects/{id}:
    put:
      summary: /subjects/{id}
      deprecated: false
      description: ''
      tags:
        - Classroom
      parameters:
        - name: id
          in: path
          description: ''
          required: true
          example: ''
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema: *ref_26
            example: ''
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - *ref_5
                  - properties:
                      subject: *ref_25
                    type: object
      security: []
  /timetables:
    get:
      summary: /timetables
      deprecated: false
      description: ''
      tags:
        - Classroom
      parameters: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - *ref_5
                  - properties:
                      timetables:
                        type: array
                        items:
                          $ref: '#/components/schemas/TimetableList'
                    type: object
      security: []
    post:
      summary: /timetables
      deprecated: false
      description: ''
      tags:
        - Classroom
      parameters: []
      requestBody:
        content:
          application/json:
            schema: &ref_28
              $ref: '#/components/schemas/TimetableCreate'
            example: ''
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - *ref_5
                  - properties:
                      timetable: &ref_27
                        $ref: '#/components/schemas/Timetable'
                    type: object
      security: []
    delete:
      summary: /timetables
      deprecated: false
      description: ''
      tags:
        - Classroom
      parameters:
        - name: ids
          in: query
          description: ''
          required: false
          schema:
            type: array
            items:
              type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - *ref_5
                  - properties:
                      data:
                        type: object
                        properties:
                          classrooms:
                            type: array
                            items: *ref_27
                    type: object
      security: []
  /timetables/{id}:
    put:
      summary: /timetables/{id}
      deprecated: false
      description: ''
      tags:
        - Classroom
      parameters:
        - name: id
          in: path
          description: ''
          required: true
          example: ''
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema: *ref_28
            example: ''
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - *ref_5
                  - properties:
                      data:
                        type: object
                        properties:
                          item: *ref_27
                    type: object
      security: []
...
  /parent/subjects:
    get:
      summary: /parent/subjects
      deprecated: false
      description: ''
      tags:
        - Diary
      parameters:
        - name: child_id
          in: query
          description: ''
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - *ref_5
                  - $ref: '#/components/schemas/ParentSubjects'
      security: []
...
components:
  schemas:
...
    TimetableCreate:
      properties:
        value:
          type: object
          examples:
            - - - 12
                - 13
                - 14
              - null
              - - 12
                - null
                - 22
                - 23
                - 23
              - null
              - - 1
                - 2
                - 3
              - null
          properties: {}
        this_week:
          type: boolean
        shift_id: *ref_0
        classroom_id: *ref_0
        school_id: *ref_0
      type: object
...
    SubjectCreate:
      properties:
        id: *ref_0
        name:
          type: string
          examples:
            - Matematika
        full_name:
          type: string
          examples:
            - Matematika baslangyc
        week_hours:
          type: integer
          examples:
            - 5
        classroom_id: *ref_0
        classroom_type: *ref_39
        classroom_type_key:
          type: integer
          examples:
            - 1
        teacher_id: *ref_0
        second_teacher_id: *ref_0
        school_id: *ref_0
        exam: *ref_23
      type: object
...
    DashboardSubjectPercent:
      properties:
        subject_id: *ref_0
        classroom_id: *ref_0
        subject_name:
          type: string
          examples:
            - Yazuw
        classroom_name:
          type: string
          examples:
            - 1A
        is_grades_full:
          type: boolean
        grades_full_percent:
          type: integer
        students_count:
          type: integer
        lessons_count:
          type: integer
        grades_count:
          type: integer
        absents_count:
          type: integer
      type: object
...
    Timetable:
      properties:
        id: *ref_0
        value:
          type: object
          examples:
            - - - 12
                - 13
                - 14
              - null
              - - 12
                - null
                - 22
                - 23
                - 23
              - null
              - - 1
                - 2
                - 3
              - null
          properties: {}
        shift: *ref_29
        classroom: *ref_20
        school: *ref_14
        updated_by: *ref_2
        created_at: *ref_4
        updated_at: *ref_4
      type: object
    ParentSubjects:
      properties:
        period_counts:
          type: array
          items:
            properties:
              period_key:
                type: integer
              lessons_count:
                type: integer
              absents_yes_count:
                type: integer
              absents_no_count:
                type: integer
              absents_ill_count:
                type: integer
            type: object
        subjects:
          type: array
          items:
            properties:
              subject: *ref_25
              period_grades:
                type: object
                properties:
                  '1': &ref_52
                    $ref: '#/components/schemas/PeriodGrade'
              exam_grade: *ref_52
              final_grade: *ref_52
            type: object
      type: object
...
    TeacherTimetable:
      properties:
        week:
          type: array
          items:
            type: array
            items:
              properties:
                date: *ref_42
                shift_times:
                  type: array
                  items:
                    type: string
                  examples:
                    - - '08:30'
                      - '09:15'
                subject: *ref_54
                subject_percent:
                  $ref: '#/components/schemas/DashboardSubjectPercent'
              type: object
      type: object
...
    PeriodGrade:
      properties:
        period_id: *ref_0
        period_key: *ref_0
        student_id: *ref_0
        grade_value:
          type: string
          examples:
            - '4.5'
        updated_at: *ref_4
      type: object
...
    Subject:
      properties:
        id: *ref_0
        name:
          type: string
          examples:
            - Matematika
        full_name:
          type: string
          examples:
            - Matematika baslangyc
        week_hours:
          type: integer
          examples:
            - 5
        classroom_type: *ref_39
        classroom_type_key:
          type: integer
          examples:
            - 1
        classroom: *ref_53
        teacher: *ref_9
        second_teacher: *ref_9
        school: *ref_50
        exam: *ref_22
      type: object
...
  securitySchemes: {}
servers: []